Add code needed to build marbl_in with POP2%ECO
1) Added MARBL_scripts/ so that buildcpp and buildnml can get necessary
   information from MARBL
2) Added manage_externals tool, and set it up to clone marbl0.34.0 into
   externals/MARBL
3) bld/build-namelist now passes necessary arguments to ocn.ecosys.tavg.sh
4) Added C1850ECO_HR and G1850ECO_HR compsets. They are identical to the
   non-_HR versions, except use SWAV instead of WW3

This commit lets us build / run compsets using POP%ECO, but there are no calls
to MARBL (the old version of BEC available in ecosys_* is used instead).

Also note that manage_externals is bringing in CVMix, but still using the code
in source/cvmix -- the only difference between that directory and
externals/CVMix/src/shared is in makedep.py (want python3 in shebang line
instead of just python).